Adding explicit declaration to use greater memory
Discovered that datasets page would not load on local39training.org server when >14K rows per user.  Signs pointed to memory allocation, but simply changing the value in php.ini did not solve the issue.  Error logs indicated the problem was in two scripts and google had a solution to increase the memory allocation for those scripts explicitly.  Change fixed the issue, committing back to master.